requirements.txt
python-arango
requests
beautifulsoup4
flask
connexion[swagger-ui]==2.7.0
